% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_dm_table.R
\name{character_dm_table}
\alias{character_dm_table}
\title{Outputs basic statistics for categorical variables}
\usage{
character_dm_table(data, var, digit)
}
\arguments{
\item{data}{The input data}

\item{var}{The categorical variable}

\item{digit}{The number of decimal places}
}
\value{
A data frame containing the mean, standard deviation, median, minimum, maximum values for the categorical variable
}
\description{
Outputs basic statistics for categorical variables
}
