% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_dm_table.R
\name{num_dm_table}
\alias{num_dm_table}
\title{Outputs basic statistics for continuous variables}
\usage{
num_dm_table(data, var, digit)
}
\arguments{
\item{data}{The input data}

\item{var}{The continuous variable}

\item{digit}{The number of decimal places}
}
\value{
A data frame containing the mean, standard deviation, median, minimum, maximum values for the continuous variable
}
\description{
Outputs basic statistics for continuous variables
}
